Small refactorings

+++ src/biodwh2-clinicaltrialsgov/src/main/java/de/unibi/agbi/biodwh2/clinicaltrialsgov/model/EventCountsStruct.java
@@ -3,17 +3,15 @@ package de.unibi.agbi.biodwh2.clinicaltrialsgov.model;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlText;
 
-import java.math.BigInteger;
-
 public class EventCountsStruct {
     @JacksonXmlText
     public String value;
     @JacksonXmlProperty(isAttribute = true, localName = "group_id")
     public String groupId;
     @JacksonXmlProperty(isAttribute = true, localName = "subjects_affected")
-    public BigInteger subjectsAffected;
+    public Integer subjectsAffected;
     @JacksonXmlProperty(isAttribute = true, localName = "subjects_at_risk")
-    public BigInteger subjectsAtRisk;
+    public Integer subjectsAtRisk;
     @JacksonXmlProperty(isAttribute = true)
-    public BigInteger events;
+    public Integer events;
 }
